requirements.txt
google-cloud-aiplatform
Pillow
mss
vertexai
flask
gunicorn